VIH-8884 user creation and notifications (#929)
Moved the responsibility of confirming the hearing to the booking based on the participant's role. Contact Email is used to identify the participant than the user name uniquely. Creat user and sending notifications responsibility is moved to the queue subscriber based on the events published by bookings API. Tidied up some of the acceptance tests so they depend o the status/values on the UI